Student Dashboard improved , board component added
